+++ MarioMe/Game.h
@@ -45,7 +45,7 @@ public:
 
 	void Init(HWND hWnd);
 	void Draw(float x, float y, D3DXVECTOR3 pivot, LPDIRECT3DTEXTURE9 texture,
-		int left, int top, int right, int bottom, int alpha = 255, int flip = 1);
+		float left, float top, float right, float bottom, int alpha = 255, int flip = 1);
 
 	void Load(string gameFile);
 	void LoadTextures(TiXmlElement* gameResources);
